Removed Base Organ by moving functionality into previously derived organs
