--- Models/Plant/Organs/RootSWIM.cs
@@ -1,123 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Text;
-using Models.Core;
-using Models.PMF.Interfaces;
-using APSIM.Shared.Utilities;
-using Models.PMF.Library;
-
-namespace Models.PMF.Organs
-{
-    /// <summary>
-    /// A root model for SWIM
-    /// </summary>
-    [Serializable]
-    public class RootSWIM : BaseOrgan, IArbitration
-    {
-        /// <summary>The uptake</summary>
-        private double[] Uptake = null;
-        /// <summary>The RLV</summary>
-        public double[] rlv = null;
-
-        private Biomass Live = new Biomass();
-        private Biomass Dead = new Biomass();
-
-        /// <summary>Link to biomass removal model</summary>
-        [ChildLink]
-        public BiomassRemoval biomassRemovalModel = null;
-
-        /// <summary>Gets or sets the water uptake.</summary>
-        /// <value>The water uptake.</value>
-        [Units("mm")]
-        public double WaterUptake
-        {
-            get { return -MathUtilities.Sum(Uptake); }
-        }
-
-        /// <summary>Gets the total biomass</summary>
-        public Biomass Total { get { return Live + Dead; } }
-
-        /// <summary>Gets the total grain weight</summary>
-        [Units("g/m2")]
-        public double Wt { get { return Total.Wt; } }
-
-        /// <summary>Gets the total grain N</summary>
-        [Units("g/m2")]
-        public double N { get { return Total.N; } }
-
-        /// <summary>Called when [water uptakes calculated].</summary>
-        /// <param name="Uptakes">The uptakes.</param>
-        [EventSubscribe("WaterUptakesCalculated")]
-        private void OnWaterUptakesCalculated(WaterUptakesCalculatedType Uptakes)
-        {
-            for (int i = 0; i != Uptakes.Uptakes.Length; i++)
-            {
-                if (Uptakes.Uptakes[i].Name == Plant.Name)
-                    Uptake = Uptakes.Uptakes[i].Amount;
-            }
-        }
-
-        /// <summary>Called when [simulation commencing].</summary>
-        /// <param name="sender">The sender.</param>
-        /// <param name="e">The <see cref="EventArgs"/> instance containing the event data.</param>
-        [EventSubscribe("Commencing")]
-        private new void OnSimulationCommencing(object sender, EventArgs e)
-        {
-            base.OnSimulationCommencing(sender, e);
-            Clear();
-        }
-
-        /// <summary>Called when crop is ending</summary>
-        /// <param name="sender">The sender.</param>
-        /// <param name="data">The <see cref="EventArgs"/> instance containing the event data.</param>
-        [EventSubscribe("PlantSowing")]
-        private void OnPlantSowing(object sender, SowPlant2Type data)
-        {
-            if (data.Plant == Plant)
-                Clear();
-        }
-
-        /// <summary>Called when crop is ending</summary>
-        /// <param name="sender">The sender.</param>
-        /// <param name="e">The <see cref="EventArgs"/> instance containing the event data.</param>
-        [EventSubscribe("PlantEnding")]
-        private void OnPlantEnding(object sender, EventArgs e)
-        {
-            Biomass total = Live + Dead;
-            if (total.Wt > 0.0)
-            {
-                Detached.Add(Live);
-                Detached.Add(Dead);
-                SurfaceOrganicMatter.Add(total.Wt * 10, total.N * 10, 0, Plant.CropType, Name);
-            }
-
-            Clear();
-        }
-
-        /// <summary>Removes biomass from organs when harvest, graze or cut events are called.</summary>
-        /// <param name="biomassRemoveType">Name of event that triggered this biomass remove call.</param>
-        /// <param name="value">The fractions of biomass to remove</param>
-        public void RemoveBiomass(string biomassRemoveType, OrganBiomassRemovalType value)
-        {
-            biomassRemovalModel.RemoveBiomass(biomassRemoveType, value, Live, Dead, Removed, Detached);
-        }
-        /// <summary>Remove maintenance respiration from live component of organs.</summary>
-        /// <param name="respiration">The respiration to remove</param>
-        public virtual void RemoveMaintenanceRespiration(double respiration)
-        {
-            double total = Live.MetabolicWt + Live.StorageWt;
-            if (respiration > total)
-            {
-                throw new Exception("Respiration is more than total biomass of metabolic and storage in live component.");
-            }
-            Live.MetabolicWt = Live.MetabolicWt - (respiration * Live.MetabolicWt / total);
-            Live.StorageWt = Live.StorageWt - (respiration * Live.StorageWt / total);
-        }
-        /// <summary>Clears this instance.</summary>
-        private void Clear()
-        {
-            Live.Clear();
-            Dead.Clear();
-        }
-    }
-}
